fix(link): fix es lint (2)
