+++ openaev-front/src/admin/components/scenarios/scenario/injects/ScenarioInjectCreation.tsx
@@ -25,8 +25,8 @@ import useDataLoader from '../../../../../utils/hooks/useDataLoader';
 import { ArticleContext, ChallengeContext, InjectContext, TeamContext } from '../../../common/Context';
 import InjectContractPicker from '../../../common/injects/create/InjectContractPicker';
 import InjectCreationConfig from '../../../common/injects/create/InjectCreationConfig';
+import { buildScenarioVariablesConfigurationUrl } from '../../ScenarioConfigurationTab';
 import articleContextForScenario from '../articles/articleContextForScenario';
-import { buildScenarioVariablesConfigurationUrl } from '../ScenarioConfiguration';
 import teamContextForScenario from '../teams/teamContextForScenario';
 
 // Full-page inject creation for scenarios: the Threat-Arsenal-style contract
